requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.29
sqlmodel==0.0.16
